sem6/iot/group-project/raspberry_pi/requirements.txt
astroid==2.3.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.7
cycler==0.10.0
firebase==3.0.1
firebase-admin==4.5.3
gcloud==0.18.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
httplib2==0.19.1
idna==2.10
isort==4.3.21
jwcrypto==0.8
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
MouseInfo==0.1.2
msgpack==1.0.2
numpy==1.17.4
oauth2client==4.1.3
packaging==20.9
Pillow==6.2.1
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.48
pycparser==2.20
pycryptodome==3.10.1
PyGetWindow==0.0.8
pylint==2.4.4
PyMsgBox==1.0.7
pyparsing==2.4.5
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.25
python-dateutil==2.8.1
python-jwt==3.3.0
PyTweening==1.0.3
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.13.0
sortedcontainers==2.1.0
sseclient==0.0.27
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.11.2
